+++ confidential-data-hub/hub/src/bin/cdh-oneshot.rs
@@ -9,9 +9,8 @@
 
 use base64::{engine::general_purpose::STANDARD, Engine};
 use clap::{Args, Parser, Subcommand};
-use confidential_data_hub::{hub::Hub, CdhConfig, DataHub};
+use confidential_data_hub::{hub::Hub, storage::volume_type::Storage, CdhConfig, DataHub};
 use log::warn;
-use storage::volume_type::Storage;
 
 #[derive(Parser)]
 #[command(name = "cdh_oneshot")]
