+++ confidential-data-hub/hub/src/error.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-use crate::image;
+use crate::{image, secret, storage};
 use thiserror::Error;
 
 pub type Result<T> = std::result::Result<T, Error>;
