+++ confidential-data-hub/hub/src/lib.rs
@@ -17,3 +17,5 @@ pub mod config;
 pub use config::*;
 
 pub mod image;
+pub mod secret;
+pub mod storage;
