+++ confidential-data-hub/hub/src/secret/error.rs
