+++ confidential-data-hub/hub/src/secret/layout/envelope.rs
