--- confidential-data-hub/secret/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "secret"
-version = "0.1.0"
-authors = ["The Confidential Container Authors"]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "secret_cli"
-required-features = [ "cli" ]
-
-[dependencies]
-anyhow.workspace = true
-async-trait.workspace = true
-base64.workspace = true
-clap = { workspace = true, optional = true }
-crypto.path = "../../attestation-agent/deps/crypto"
-kms = { path = "../kms", default-features = false }
-rand = { workspace = true, optional = true }
-serde = "1"
-serde_json = "1"
-strum = { workspace = true, features = ["derive"] }
-thiserror.workspace = true
-tokio = { workspace = true, features = ["io-util"], optional = true }
-zeroize.workspace = true
-
-[dev-dependencies]
-assert_cmd = "2.0"
-assert-json-diff.workspace = true
-rstest.workspace = true
-tempfile.workspace = true
-
-[features]
-default = [ "cli" ]
-cli = ["clap/derive", "rand", "tokio/rt-multi-thread", "tokio/sync", "tokio/macros"]
-
-aliyun = ["kms/aliyun"]
-kbs = ["kms/kbs"]
-sev = ["kms/sev"]
-ehsm = ["kms/ehsm"]
