--- confidential-data-hub/storage/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "storage"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow.workspace = true
-async-trait.workspace = true
-base64.workspace = true
-log.workspace = true
-kms = { path = "../kms", features = ["kbs"] }
-rand = { workspace = true, optional = true }
-secret = { path = "../secret" }
-serde.workspace = true
-serde_json.workspace = true
-strum = { workspace = true, features = ["derive"] }
-tempfile = { workspace = true, optional = true }
-thiserror.workspace = true
-tokio = { workspace = true, optional = true }
-
-[dev-dependencies]
-rstest.workspace = true
-tokio = { workspace = true, features = ["rt", "macros" ] }
-
-[build-dependencies]
-anyhow.workspace = true
-
-[features]
-default = ["aliyun"]
-aliyun = [ "rand", "tempfile", "tokio/fs", "tokio/process", "tokio/io-util", "tokio/time" ]
